refactor: GameController.js 파일을 Controller 폴더로 이동
